Move infrahouse repo configuration into bootcmd
When the AMI already includes the infrahouse APT list file, cloud-init add a duplicate entry and that breaks instance provisioning.
Therefore, the bootcmd script checks if the infrahouse.list file is already installed. If not, it adds it.

Now, because the InfraHouse repo is surely present, the puppet section in the cloud-init is redundant. It actually fails - so removing it.

Added test `tests/test_single_instance.py`. This is to ensure the instance can actually start with the generated user-data.