Move infrahouse repo configuration into bootcmd (#43)
* Move infrahouse repo configuration into bootcmd

When the AMI already includes the infrahouse APT list file, cloud-init add a duplicate entry and that breaks instance provisioning.
Therefore, the bootcmd script checks if the infrahouse.list file is already installed. If not, it adds it.

Now, because the InfraHouse repo is surely present, the puppet section in the cloud-init is redundant. It actually fails - so removing it.

Added test `tests/test_single_instance.py`. This is to ensure the instance can actually start with the generated user-data.

* Fixing formatting

* Remove unused puppet_version_map

* Move to infrahouse-core dependency

* Remove tflint

I'm not going to let go my emotions, will just say.
Not providing command line arguments, or packages, or any other way but requiring root unconditionally for a tool that just reads text - is an interesting design choice.

* And remove tflint from linters

* Fix imports after moving to pytest-infrahouse

* Add python-mimeparse for YAML parsing

* For now, only us-west-1 is supported